feat: move config to dedicated package.
logion-network/logion-internal#1283
